+++ CHANGELOG.md
@@ -42,6 +42,8 @@
 - improve compatibility with providers changing the Message-ID
   (as Outlook.com) #2250 #2265
 
+- correctly show emails that were sent to an alias and then bounced 
+
 - implement Consistent Color Generation (XEP-0392),
   that results in contact colors be be changed #2228 #2229 #2239